My most ambitious project as of April 2025
A sleep stage classification project using a hybrid CNN+LSTM Model. Achieved 96% accuracy on the Sleep EDF Dataset(using 10 files(10psg+10hypnograms) as the dataset was too large to be processed. Hypnograms consist of the labels of sleep stages. The project implements usage of various signals together instead of just 1 as of other available projects in this domain. The signals used are: EEG(brain signals), EOG(eye movement), EMG(muscle activity), and respiratory signals. This leads to accuracy in the higher levels as compared to related work in this domain